refactor: move badge logic to gamification domain
