+++ apps/govquests-frontend/src/domains/gamification/components/BadgeCard.tsx
@@ -1,7 +1,7 @@
 import { cn } from "@/lib/utils";
 import Image from "next/image";
 import Link from "next/link";
-import { useFetchBadge } from "../hooks/useFetchBadge";
+import { useFetchBadge } from "../../questing/hooks/useFetchBadge";
 
 interface BadgeCardProps {
   badgeId: string;
