+++ apps/govquests-frontend/src/domains/gamification/graphql/badgeQuery.ts
@@ -1,0 +1,14 @@
+import { graphql } from "gql.tada";
+
+export const BadgeQuery = graphql(`
+  query GetBadge($id: ID!) {
+    badge(id: $id) {
+      id
+      displayData {
+        title
+        description
+        imageUrl
+      }
+    }
+  }
+`);
