+++ apps/govquests-frontend/src/domains/gamification/hooks/useFetchBadge.ts
@@ -1,0 +1,12 @@
+"use client";
+
+import { useQuery } from "@tanstack/react-query";
+import { fetchBadge } from "../services/badgeService";
+
+export const useFetchBadge = (id: string) => {
+  return useQuery({
+    queryKey: ["badge", id],
+    queryFn: () => fetchBadge(id),
+    enabled: !!id,
+  });
+};
