+++ apps/govquests-frontend/src/domains/gamification/services/badgeService.ts
@@ -1,0 +1,7 @@
+import { API_URL } from "@/lib/utils";
+import request from "graphql-request";
+import { BadgeQuery } from "../graphql/badgeQuery";
+
+export const fetchBadge = async (id: string) => {
+  return await request(API_URL, BadgeQuery, { id });
+};
