+++ apps/govquests-frontend/src/domains/gamification/types/badgeTypes.ts
@@ -1,0 +1,4 @@
+import { ResultOf } from "gql.tada";
+import { BadgeQuery } from "../graphql/badgeQuery";
+
+export type Badge = ResultOf<typeof BadgeQuery>["badge"];
